requirements.txt
appdirs==1.4.3
autopep8==1.4.4
beautifulsoup4==4.8.2
bootstrap4==0.1.0
certifi==2019.11.28
chardet==3.0.4
config==0.5.0.post0
confusable-homoglyphs==3.2.0
distlib==0.3.0
dj-database-url==0.5.0
Django==1.11
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-contrib-comments==1.9.1
django-crispy-forms==1.7.2
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-likes==2.0.1
django-registration==2.4.1
django-secretballot==1.0.0
django-tinymce==2.8.0
filelock==3.0.12
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.2.0
Pillow==7.0.0
pipenv==2018.11.26
psycopg2==2.7.5
pycodestyle==2.5.0
python-dateutil==1.5
python-decouple==3.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.9
virtualenv==20.0.8
virtualenv-clone==0.5.3
whitenoise==4.1.3
zipp==3.1.0
